[JENKINS-42746] - Make the plugin independent from the XML sample in the core
In Jenkins 2.50 we have got JENKINS-42724, which has been caused by the usage of the old resource file located within the Jenkins core.
It is likely a leftover from the migration of the windows-slaves-installer module.

Since the Windows Slaves plugin has its internal macro resolution logic, the sample should be also inside plugin. A longer-term fix is to make the plugin depent on the Windows Agent Installer module