Remove CHANGELOG from current branch
Make it clear to readers that they can see the version history in GitHub
releases, not in the changelog file.  The changelog file has not been
changed since release 1.5, so the README reference to the changelog can
use the 1.5 tag to show the historical version without distracting the
reader of the current branch.